+++ packages/lumx-core/src/js/types/jsx/PropsToOverride.ts
@@ -7,4 +7,6 @@ export type PropsToOverride =
     | 'handleClose'
     | 'handleBeforeClick'
     | 'handleAfterClick'
+    | 'handleMouseEnter'
+    | 'handleMouseLeave'
     | 'handleKeyDown';